+++ backend/canisters/neuron_controller/CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 
 - Separate spawning neurons from active neurons ([#5159](https://github.com/open-chat-labs/open-chat/pull/5159))
+- Refresh neurons after spawning or disbursing ([#5160](https://github.com/open-chat-labs/open-chat/pull/5160))
 
 ## [[2.0.1002](https://github.com/open-chat-labs/open-chat/releases/tag/v2.0.1002-neuron_controller)] - 2024-01-10